# ARROW-13965: [C++] dynamic_casts in parquet TypedColumnWriterImpl impacting performance