--- a/cpp/src/parquet/column_writer.cc
+++ b/cpp/src/parquet/column_writer.cc
@@ -1163,9 +1163,8 @@ class TypedColumnWriterImpl : public ColumnWriterImpl, public TypedColumnWriter<
 
   void WriteDictionaryPage() override {
     DCHECK(current_dict_encoder_);
-    std::shared_ptr<ResizableBuffer> buffer =
-        AllocateBuffer(properties_->memory_pool(),
-                       current_dict_encoder_->dict_encoded_size());
+    std::shared_ptr<ResizableBuffer> buffer = AllocateBuffer(
+        properties_->memory_pool(), current_dict_encoder_->dict_encoded_size());
     current_dict_encoder_->WriteDict(buffer->mutable_data());
 
     DictionaryPage page(buffer, current_dict_encoder_->num_entries(),
@@ -1211,8 +1210,8 @@ class TypedColumnWriterImpl : public ColumnWriterImpl, public TypedColumnWriter<
   // downcasted observers of current_encoder_. perform the downcast once as opposed
   // to at every use since dynamic_cast is so expensivea, and static cast
   // is not available due to virtual inheritance
-  ValueEncoderType * current_value_encoder_;
-  DictEncoder<DType> *  current_dict_encoder_;
+  ValueEncoderType* current_value_encoder_;
+  DictEncoder<DType>* current_dict_encoder_;
   std::shared_ptr<TypedStats> page_statistics_;
   std::shared_ptr<TypedStats> chunk_statistics_;
 
@@ -1365,7 +1364,7 @@ class TypedColumnWriterImpl : public ColumnWriterImpl, public TypedColumnWriter<
       current_encoder_ = MakeEncoder(DType::type_num, Encoding::PLAIN, false, descr_,
                                      properties_->memory_pool());
       current_value_encoder_ = dynamic_cast<ValueEncoderType*>(current_encoder_.get());
-      current_dict_encoder_ = 0; // not using dict
+      current_dict_encoder_ = 0;  // not using dict
       encoding_ = Encoding::PLAIN;
     }
   }