@pitrou 2021-09-13
TBH, the stream APi is probably not high-performance at all. I'm not sure it makes sense to write benchmarks for it (unless we want to showcase how slow it is :-)).