@etseidl 2021-09-13
Right now I'm trying to modify the stream_reader_writer example to see if the speed bump is noticeable there.  If not, I'll  modify my code (which doesn't actually use the streaming api, it just writes in batches of 1) to  do writes to a fixed schema.