@etseidl 2021-09-13
Ok, I modified the stream_reader_writer example to just do the write, and changed the number of rows from 2000 to 20M (and the max row group size from 1000 to 10M).  I ran this on my laptop with and without the changes in this patch.  I used the csh time command to compare run times. The results are:
```
original:
52.919u 2.146s 0:55.81 98.6%	0+0k 0+2646928io 0pf+0w
53.328u 2.327s 0:56.38 98.6%	0+0k 0+2646928io 0pf+0w
53.048u 2.341s 0:56.06 98.7%	0+0k 0+2646928io 0pf+0w

patched:
35.367u 2.012s 0:37.88 98.6%	0+0k 0+2646928io 0pf+0w
35.133u 1.717s 0:37.36 98.6%	0+0k 0+2646928io 0pf+0w
35.228u 1.950s 0:37.61 98.8%	0+0k 0+2646928io 0pf+0w
```
@emkornfield Do you want benchmark code added to the PR, or are these results sufficient?