@pitrou 2021-09-13
As far as am I concerned (and not wanting to speak for @emkornfield  :-)), I would be ok with simply adding these numbers to the PR description.