@emkornfield 2021-09-13
Yes, I think just updating the description is fine.