@pitrou 2021-09-16
Note the LZ4 build failures on MinGW are unrelated.